Доработал проект по MVC v.2
