+++ app/src/main/java/com/example/moviesearch/ui/poster/PosterActivity.kt
@@ -2,22 +2,16 @@ package com.example.moviesearch.ui.poster
 
 import android.app.Activity
 import android.os.Bundle
-import android.widget.ImageView
-import com.bumptech.glide.Glide
+import com.example.moviesearch.Creator
 import com.example.moviesearch.R
 
 class PosterActivity : Activity() {
 
-    private lateinit var poster: ImageView
+    private val posterController = Creator.providePosterController(this)
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_poster)
-        poster = findViewById(R.id.poster)
-        val url = intent.extras?.getString("poster", "")
-
-        Glide.with(applicationContext)
-            .load(url)
-            .into(poster)
+        posterController.onCreate()
     }
 }
